requirements.txt
requests>=2.0
beautifulsoup4>=4.0
Pillow>=8.0
